refactor: update encryptionType filename
