+++ libs/common/src/models/domain/enc-array-buffer.ts
@@ -1,4 +1,4 @@
-import { EncryptionType } from "../../enums/encryptionType";
+import { EncryptionType } from "../../enums/encryption-type";
 import { IEncrypted } from "../../interfaces/IEncrypted";
 import { Utils } from "../../misc/utils";
