+++ libs/common/src/services/crypto.service.ts
@@ -12,7 +12,7 @@ import { ProfileOrganizationResponse } from "../admin-console/models/response/pr
 import { ProfileProviderOrganizationResponse } from "../admin-console/models/response/profile-provider-organization.response";
 import { ProfileProviderResponse } from "../admin-console/models/response/profile-provider.response";
 import { KdfConfig } from "../auth/models/domain/kdf-config";
-import { EncryptionType } from "../enums/encryptionType";
+import { EncryptionType } from "../enums/encryption-type";
 import { HashPurpose } from "../enums/hashPurpose";
 import {
   DEFAULT_ARGON2_ITERATIONS,
